split routes file
