+++ backend/app/controllers/calendar/fetchCalendars.test.js
@@ -1,24 +1,4 @@
-const { request } = require( '../tests/setup.js' );
-
-it( 'Does not provide unnecessary endpoints', async () => {
-   const res = await request.get( '/' );
-   expect( res.statusCode ).toBe( 404 );
-
-   const res2 = await request.get( '/api' );
-   expect( res2.statusCode ).toBe( 404 );
-
-   const res3 = await request.get( '/api/calendar' );
-   expect( res3.statusCode ).toBe( 404 );
-
-   const res4 = await request.get( '/api/calendar/' );
-   expect( res4.statusCode ).toBe( 404 );
-} );
-
-it( 'Provides /api/calendars endpoint', async () => {
-   const res = await request.get( '/api/calendars' );
-   expect( res.statusCode ).toBe( 200 );
-   expect( Array.isArray( res.body ) ).toBe( true );
-} );
+const { request } = require( '../../../tests/setup.js' );
 
 it( 'Returns calendar via /api/calendar/*', async () => {
    const res = await request.get( '/api/calendar/test?startDate=2020-01-01' );
